HOM-201 P1-2: move <img> rewrite from render layer to IO layer
为何
- ReadmeWebView.loadIfNeeded 每次切 repo 都对 htmlFragment 跑一次
  NSRegularExpression 全文扫描重写 <img> 相对路径,大 README(react /
  kubernetes 等 300-500KB HTML)每切一次都重新算 —— 渲染线程上跑既浪费
  又延迟首帧。

改动
- 新建 ReadmeAssetURLRewriter(`Core/Network/GitHubAPI/`):enum namespace +
  rewrite(in:owner:repo:) / rewriteOne(_:rawBase:),逻辑从 ReadmeWebView
  搬过来零修改;
- ReadmeAPI 在 4 个 200 / 兜底分支(performRefreshReadme /
  performRefreshTrendingReadme / refreshUnconditional /
  refreshTrendingUnconditional)upsert 前调 rewrite,落到 readmes.rendered_html
  / trending_readmes.rendered_html 的就是 rewrite 后版本;size 同步用
  rewrittenHtml.utf8.count(与磁盘实占对齐,LRU 淘汰更准确);
- ReadmeWebView:删 owner / repo 两参数;loadIfNeeded 内直接 assembleDocument
  不再跑正则;删 rewriteAssetURLs / rewriteOneAssetURL 两 static 方法;
- ReadmeStateView(RepoDetailView)同步删 owner / repo 两参数;
- 4 个 DetailContent(Manage / Trending / Weekly / ActivityRepo)调
  ReadmeStateView 时去掉 owner/repo 传参;
- 测试:原 ReadmeWebViewTests 关于 rewrite 的 11 项搬到新文件
  ReadmeAssetURLRewriterTests(namespace 改成 ReadmeAssetURLRewriter,
  逻辑零改动);ReadmeWebViewTests 只保留 baseURL 1 项;
- 加 ReadmeAPINetworkTests 2 项 rewrite-at-upsert 验证(manage / trending
  各一)。

收益
- 详情切换 0 正则成本(大 README 切换不再延迟首帧);
- 每条 README 一辈子只 rewrite 一次(在 IO 线程 + 落库前);
- ReadmeWebView API 收窄,UI 层无需 owner/repo 上下文。

约束
- 项目未上线,旧数据不迁移,新跑 refresh / 200 时自动覆盖;
- baseURL 仍保留供链接 <a href> 解析(GitHub render 的链接仍是相对路径,
  本步不在范围);
- promote 路径(trending → manage)不二次 rewrite,因 trending 表落库时已是
  rewrite 后版本。

环境局限
- 当前环境无完整 Xcode,未跑 xcodebuild;改动是纯结构性迁移 + API 收窄 +
  测试同款搬移,逻辑零改动。
