+++ docs/工程进度/功能实现总览.md
@@ -1113,6 +1113,7 @@ W4-4 Batch C+D 一次落地 7 commit（C1 Rate Limit 退避 / C2 ETag 缓存 / C
 
 | 日期 | 变更 |
 |---|---|
+| 2026-06-14 傍晚 | ⚡ **HOM-201 P1-2：`<img>` 相对路径重写从渲染层迁移到 IO 层**：原 `ReadmeWebView.loadIfNeeded` 每次切 repo 都对 `htmlFragment` 跑一次 `NSRegularExpression` 全文扫描重写 `<img>` 相对路径——大 README（react / kubernetes 这种 300-500KB HTML）每切一次都重新算,渲染线程上跑既浪费又延迟首帧。**改动**：① 新建 `Starcat/Core/Network/GitHubAPI/ReadmeAssetURLRewriter.swift`（`enum` namespace + 2 个 static 方法 `rewrite(in:owner:repo:)` / `rewriteOne(_:rawBase:)`），逻辑从 `ReadmeWebView` 搬过来,**零修改**;② `ReadmeAPI` 在 4 个 200 / 兜底分支（`performRefreshReadme` / `performRefreshTrendingReadme` / `refreshUnconditional` / `refreshTrendingUnconditional`）upsert 前调 `ReadmeAssetURLRewriter.rewrite(...)`，落到 `readmes.rendered_html` / `trending_readmes.rendered_html` 的就是 rewrite 后的版本；`size` 同步用 `rewrittenHtml.utf8.count`（与磁盘实占对齐,后续 LRU 淘汰按 size 判断更准确）；③ `ReadmeWebView`：删 `owner: String?` / `repo: String?` 两个参数,`loadIfNeeded` 内删去 `Self.rewriteAssetURLs(...)` 调用直接 `assembleDocument(fragment: htmlFragment, ...)`,删 `static rewriteAssetURLs` / `rewriteOneAssetURL` 两个方法；④ `ReadmeStateView`（`RepoDetailView.swift`）同步删 `owner` / `repo` 两参数；⑤ 4 个 DetailContent（`ManageDetailContent` / `TrendingDetailContent` / `WeeklyDetailContent` / `ActivityRepoDetailContent`）调 `ReadmeStateView` 时去掉 owner/repo 传参；⑥ 测试拆分：原 `ReadmeWebViewTests` 关于 rewrite 的 11 项用例搬到新文件 `ReadmeAssetURLRewriterTests`（namespace 改 `ReadmeAssetURLRewriter.rewrite` / `rewriteOne`,逻辑零改动）,`ReadmeWebViewTests` 只保留 baseURL 1 项；⑦ `ReadmeAPINetworkTests` 新增 2 项 rewrite-at-upsert 验证：refreshReadme / refreshTrendingReadme 200 时落库的 rendered_html 已 rewrite。**收益**：① 详情切换 0 正则成本（大 README 切换不再延迟首帧）；② 每条 README 一辈子只 rewrite 一次（在 IO 线程 + 落库前）；③ ReadmeWebView API 收窄,UI 层无需 owner/repo 上下文。**约束**：① 项目未上线,旧数据不迁移,新跑 refresh / 200 时自动覆盖；② `baseURL` 仍保留供链接 `<a href>` 解析（GitHub render 出来的链接仍是相对路径,这一份不在本步范围）；③ promote 路径（trending → manage）不二次 rewrite,因 trending 表落库时已是 rewrite 后版本；④ 当前环境无完整 Xcode,未跑 `xcodebuild`,改动是结构性迁移 + API 收窄 + 测试同款搬移。涉及 10 文件（新 2 + 改 7 + 测试拆分 1）。 |
 | 2026-06-14 傍晚 | ⚡ **HOM-201 P1-1：列表 hover prefetch README**：详情页打开瞬间默认要走一帧 `.loading`，命中缓存也得 GRDB 异步读一次；把"读缓存 + 必要时 304 验证"提前到列表 hover 期间，用户实际点击进入时缓存已就绪。**改动**：① 新建 `Starcat/Shared/Components/ReadmePrefetchModifier.swift`：`@State Task` 持有当前 in-flight hover task + 500ms debounce + `onHover` 离开取消（防抖、防快速划过批量请求）；提供 `.readmePrefetch { ... }` View modifier；② `ReadmeAPI` 加 `prefetch(for: Repo)` 与 `prefetchTrending(owner: String, repo: String)` —— 都先用 `isWithinSoftTtl(softTtl=6h)` 短路本地 fresh 缓存（命中直接 return 不打 GitHub），未命中或过期才走 `refreshReadme` / `refreshTrendingReadme`（被 `ReadmeInflightTracker` 并发去重）；③ `RepoListView`（manage）/`TrendingView`（trending）/`WeeklyContentView`（weekly）/`ActivityView`（active）各列表的 `UnifiedRepoRow` 都挂上 `.readmePrefetch`：manage / active 调 `prefetch(for:)` 命中 `readmes` 表，trending / weekly 调 `prefetchTrending(...)` 命中 `trending_readmes` 表；④ `ReadmeAPINetworkTests` 新增 5 项 prefetch 测试覆盖"fresh 短路 / 过期触发 refresh / 无缓存触发 refresh"3 个分支 × manage + trending 路径。**收益**：详情秒开率显著提升；softTtl 6h 内的 hover **完全不发 GitHub 请求**；P0-3 inflight 去重保证多窗口并发 hover 同一 repo 也只发一次。**约束**：① trending 路径 hover prefetch 在 P1-4 前会出现"prefetch 走 304 验证 / 详情页 loadTrending 又强制 refresh"的边缘 case（多一次 304 请求），P1-4 让 loadTrending 也用 softTtl 后闭环；② SearchCenter 的 row 故意不挂（搜索结果用户多半只查不进详情，prefetch 边际收益小）；③ 错误静默吞掉（包括读缓存 throws / 网络失败），用户真进详情时 `loadInternal` 正常 SWR 流程会再尝试并显示错误态。涉及 7 文件（新 1 + 改 5 + 测试 1）。 |
 | 2026-06-14 下午 | ⚡ **HOM-201 P0-1：trending → manage README 缓存 promote**：用户在 trending 详情读过 README → star → 切到 manage 详情时,旧逻辑会重新打 GitHub 拉同一份内容（两表用不同 PK：manage 用 `repo_id`,trending 用 `full_name`,互相看不见）。**改动**：① `ReadmeAPI.cachedReadme(repoId:)` 签名升级为 `cachedReadme(for repo: Repo)` —— manage 表未命中时兜底查 `trending_readmes` 表,命中则 upsert 到 `readmes` 表 + 删 trending 行（**双份存储清理由本步同步完成,不留 stale**）；promote 写失败仍返回 trending 数据给 UI 用,下次再尝试；② `ReadmeViewModel.loadInternal` 改调新签名(传 repo 而非 repoId)；③ `ReadmeAPINetworkTests` 新增 2 项测试覆盖 promote 与短路（manage hit 不查 trending）。**单向 promote**：不做反向兜底（manage → trending）因为 `cachedTrendingReadme(owner:repo:)` 拿不到 `repo.id` 没法反查,且反向场景较少。**收益**：从 trending 转 star 后第一次进 manage 详情 0 GitHub 请求；trending_readmes 表逐步迁移到 readmes 表,避免双份占用。**约束**：① promote 时 `Readme.content` 设 nil(trending 表没存 raw markdown),后续 `onHTMLLoaded` backfill 流程会补；② 修改 `cachedReadme` 签名导致文档 `readme.md-渲染设计.md` / `功能实现总览.md` 历史变更日志的旧记述与新代码不一致,新增条目以本次为准。涉及 3 文件（改 2 + 测试 1）。 |
 | 2026-06-14 下午 | 🩹 **HOM-201 P0-4：active 详情页补挂 `onHTMLLoaded` 回调**：manage 路径 README 拉到后会异步补 raw markdown 落 `readmes.content` 并视情况触发向量索引重建（文档 `26-向量搜索改进.md` §6 承诺），但 active 详情页（`ActivityDetailScaffoldShell.ensureReadmeViewModel`）漏接同一个 closure——active 走 `loadInternal` 跟 manage 同源、依赖同一个 `readmes` 表，但已 star 仓库从 active 详情读 README 后既不补 markdown、也不重建向量,造成"manage 详情进 → 拉到 README 后 backfill；切到 active 同 repo → 无后续动作"的语义割裂。**改动**：① `AppDependencies` 新增工厂方法 `makeReadmeOnHTMLLoadedHandler()`（capture `readmeAPI` / `semanticSearchService` 两个 App 级长寿命引用，不持有 self），把原来分散在 `HomeView.init` 内的 inline closure 提到 DI 层单一来源；② `HomeView.init` 加 `readmeOnHTMLLoaded` 必填参数，`_readmeVM` 直接用注入的 handler 替代 inline 构造；③ `ContentView` 调用 HomeView 时传 `dependencies.makeReadmeOnHTMLLoadedHandler()`；④ `ActivityDetailScaffoldShell.ensureReadmeViewModel` 新挂同一工厂 → 补齐 active 路径的 backfill；⑤ `WeeklyDetailScaffoldShell.ensureReadmeViewModel` **故意不挂**（weekly 走 `loadTrending` 永不触发 `onHTMLLoaded`，让代码层显式表达"weekly 是 trending 路径"）。**收益**：active 详情读已 star repo README 后会自动补 raw markdown + 视情况重建语义向量；HomeView 内 backfill closure 不再重复定义。**约束**：① 工厂方法每次返回新 closure 实例（每个 VM 独立 closure，便于排查；不缓存避免持有循环）；② weekly 路径不挂，未来若 weekly 切到 `loadInternal` 需同步补挂。涉及 4 文件。 |
